# api/openapi.yaml
openapi: 3.0.3
info:
  title: Velox Billing API
  description: |
    Open-source usage-based billing engine. Handles pricing, subscriptions,
    usage metering, invoice generation, Stripe payments, dunning, and credits.
  version: "2026-04-07"
  license:
    name: MIT
    url: https://github.com/sagarsuperuser/velox/blob/main/LICENSE
  contact:
    name: Velox
    url: https://github.com/sagarsuperuser/velox
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Use `Authorization: Bearer vlx_secret_...`
        Key types: vlx_platform_ (tenant mgmt), vlx_secret_ (full access), vlx_pub_ (restricted)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, rate_limit_error, api_error]
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    InvoiceStatus:
      type: string
      enum: [draft, finalized, paid, voided, uncollectible]
      description: |
        Lifecycle state of the invoice. `uncollectible` marks an invoice
        a dunning policy has given up collecting (no further retries) while
        keeping it on the books — distinct from `voided` (cancelled, owes
        nothing).
    InvoicePaymentStatus:
      type: string
      enum: [pending, processing, succeeded, failed, unknown]
      description: |
        Payment-attempt status. `unknown` marks a Stripe charge attempt that
        returned an ambiguous error (5xx, timeout, connection reset) where
        we cannot tell whether Stripe processed the charge — a reconciler
        resolves these by querying Stripe after a cool-off window.
    InvoiceTaxStatus:
      type: string
      enum: [ok, pending, failed]
      description: |
        Whether tax has been successfully calculated. Happy path is `ok`.
        `pending` means a transient calculation failure is awaiting retry;
        `failed` means retries are exhausted (operator resolves manually).
        Pending/failed invoices are blocked from finalize so we never send
        a wrong tax amount.
...
    Attention:
      type: object
      description: |
        Unified "this invoice needs operator attention" surface.
        Computed server-side from durable invoice fields
        (`tax_status`, `tax_error_code`, `payment_status`,
        `last_payment_error`). Never persisted —
        always derived. Omitted entirely when the invoice is healthy
        (matches Stripe's `last_finalization_error: null` ergonomic).
        See ADR-009.
      required: [severity, reason, message]
      properties:
        severity:
          $ref: "#/components/schemas/AttentionSeverity"
        reason:
          $ref: "#/components/schemas/AttentionReason"
        message:
          type: string
          description: Human-readable headline copy, safe to render verbatim.
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AttentionActionItem"
          description: |
            Prescribed next steps. First item is the primary CTA; the
            rest are secondary.
        code:
          type: string
          description: |
            Open, dotted, provider-specific code (e.g.
            `tax.customer_data_invalid`, `payment.declined`).
            New codes ship without a contract
            bump. Stripe parity.
        doc_url:
          type: string
          description: |
            Deep link to operator-facing documentation for this
            reason. Stripe parity.
        param:
          type: string
          description: |
            Dotted-path pointer at the entity field the operator must
            edit to resolve the issue (e.g.
            `customer.address.postal_code`). Stripe parity. Empty
            when the resolution isn't field-scoped.
        detail:
          type: string
          description: |
            Raw provider payload (Stripe Tax JSON envelope, last
            payment error message). Disclosed in a collapsible
            section for diagnostic depth.
        since:
          type: string
          format: date-time
          description: |
            When the attention condition started — operators triage
            by age. tax_deferred_at for tax reasons; due_at for
            overdue; updated_at as a proxy for payment failures.
...
    Invoice:
      type: object
      description: |
        A Velox invoice. Returned from create, finalize, void, list, get,
        rotate-public-token, and collect-payment endpoints.
      required:
        - id
        - customer_id
        - subscription_id
        - invoice_number
        - status
        - payment_status
        - currency
        - subtotal_cents
        - discount_cents
        - tax_amount_cents
        - tax_rate
        - total_amount_cents
        - amount_due_cents
        - amount_paid_cents
        - credits_applied_cents
        - billing_period_start
        - billing_period_end
        - net_payment_term_days
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: vlx_inv_42014438d82ddf33
        tenant_id:
          type: string
          description: |
            Owning tenant. Returned for diagnostic clarity but redundant
            with the API key (which already determines the tenant).
            Omitted when empty.
        customer_id:
          type: string
        subscription_id:
          type: string
          description: Empty for one-off invoices not tied to a subscription.
        invoice_number:
          type: string
          description: Empty until the invoice is finalized.
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        payment_status:
          $ref: "#/components/schemas/InvoicePaymentStatus"
        currency:
          type: string
          example: USD
        subtotal_cents:
          type: integer
          format: int64
        discount_cents:
          type: integer
          format: int64
        tax_amount_cents:
          type: integer
          format: int64
        tax_rate:
          type: number
          format: double
          description: Tax rate as a percent (18.50 = 18.50%), up to 4-decimal precision. ADR-042/043.
        tax_name:
          type: string
        tax_country:
          type: string
        tax_id:
          type: string
        tax_provider:
          type: string
          description: |
            Durable audit snapshot of the upstream tax engine that ran the
            calculation (e.g. `stripe`, `none`, `manual`).
        tax_calculation_id:
          type: string
          description: |
            Provider-side calculation reference captured at build time.
            For Stripe Tax this expires after 24 h; persisted so finalized
            invoices remain reconstructable past expiry.
        tax_transaction_id:
          type: string
          description: |
            Provider-side committed transaction reference, captured at
            finalize. For Stripe Tax this is `tx_xxx` and is the handle
            the provider needs for a later reversal when a credit note is
            issued.
        tax_reverse_charge:
          type: boolean
        tax_exempt_reason:
          type: string
        tax_status:
          $ref: "#/components/schemas/InvoiceTaxStatus"
        tax_deferred_at:
          type: string
          format: date-time
          nullable: true
        tax_retry_count:
          type: integer
        tax_pending_reason:
          type: string
        tax_error_code:
          type: string
          description: |
            Typed taxonomy of `tax_pending_reason`. One of
            `customer_data_invalid`, `jurisdiction_unsupported`,
            `provider_outage`, `provider_auth`, `unknown`. Empty for
            invoices deferred before this column existed (migration
            0067).
        attention:
          $ref: "#/components/schemas/Attention"
        recovery_warning:
          allOf:
            - $ref: "#/components/schemas/RecoveryWarning"
          description: |
            What an operator must know BEFORE recording an OFFLINE payment on a
            written-off invoice — a consequence that will go unreconciled, NOT
            a refusal. Absent when there is nothing to say.
        total_amount_cents:
          type: integer
          format: int64
        amount_due_cents:
          type: integer
          format: int64
        amount_paid_cents:
          type: integer
          format: int64
        credits_applied_cents:
          type: integer
          format: int64
        billing_period_start:
          type: string
          format: date-time
        billing_period_end:
          type: string
          format: date-time
        billing_period_display:
          type: string
          description: >-
            Human period string with the inclusive last covered day
            ("Jun 1, 2028 – Jun 30, 2028"), date-only in the invoice's billing
            timezone (ADR-058 / ADR-077). Computed on read; the raw
            billing_period_start/end stay half-open. Empty/omitted for one-off
            invoices with no period.
        billing_timezone:
          type: string
          description: >-
            IANA timezone the period boundaries are anchored in — the org
            billing timezone resolved and denormalized onto the invoice at issue
            (ADR-077). billing_period_display is rendered in this zone, so an
            issued invoice's dates stay fixed even if the org later changes its
            timezone. Empty for ad-hoc/legacy invoices.
        issued_at:
          type: string
          format: date-time
          nullable: true
          description: Set when the invoice transitions out of draft.
        due_at:
          type: string
          format: date-time
          nullable: true
        paid_at:
          type: string
          format: date-time
          nullable: true
        voided_at:
          type: string
          format: date-time
          nullable: true
        stripe_payment_intent_id:
          type: string
        last_payment_error:
          type: string
        auto_charge_pending:
          type: boolean
        public_token:
          type: string
          description: |
            Hosted-invoice-URL credential (Stripe-parity
            `hosted_invoice_url`). Generated at finalize; drafts have an
            empty token. Rotatable via `POST
            /v1/invoices/{id}/rotate-public-token`.
        net_payment_term_days:
          type: integer
        memo:
          type: string
        footer:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        source_plan_changed_at:
          type: string
          format: date-time
          nullable: true
        source_subscription_item_id:
          type: string
        source_change_type:
          $ref: "#/components/schemas/ItemChangeType"
        billing_reason:
          $ref: "#/components/schemas/InvoiceBillingReason"
        stripe_invoice_id:
          type: string
          description: |
            Source Stripe invoice id (`in_xxx`) for an invoice imported from
            a Stripe account. Currently always empty: the Stripe importer was
            removed (deferred until a design-partner cutover) and nothing
            writes this field today. The column and its unique index remain
            for a future importer. Empty for Velox-native invoices.
        is_simulated:
          type: boolean
          description: |
            Whether this invoice's domain timestamps (created/issued/due/paid)
            were stamped on a test clock's simulated time rather than
            wall-clock. Captured at write time (engine: the subscription is
            clock-pinned; manual composer: the customer is clock-pinned). The
            dashboard renders a "simulated" badge from this authoritative flag
            instead of inferring it. Always false in live mode.
...
    InvoiceWithLineItems:
      type: object
      description: |
        Response shape for `GET /v1/invoices/{id}`. The handler returns the
        invoice and its line items as separate top-level keys (rather than
        nesting `line_items` inside the invoice) so the client can render
        the two arrays independently.
      required: [invoice, line_items]
      properties:
        invoice:
          $ref: "#/components/schemas/Invoice"
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
...
paths:
...
  /v1/invoices/{id}:
    get:
      tags: [Invoices]
      summary: Get invoice with line items
      operationId: getInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice with its line items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithLineItems"
        "404":
          description: Invoice not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/invoices/{id}/pdf:
    get:
      tags: [Invoices]
      summary: Download invoice PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/pdf: {}
...